GMPE function for Boore et al.(2014) NGA-West2 model

BSSA14 returns the ground-motion prediction with it sigma of Boore et al.(2014) GMPE.

BSSA14(Mag, Rjb, Prd, Vs30, ftype = 0, Z1.0, regionflag = 0,
  basinflag = 0)

Arguments

Mag
Earthquake momnet magnitude, Numeric.

Rjb
Joyner and Boore distance(km), Numeric.

Prd
Period of spectral acceleration.

Vs30
Vs30(m/s).

ftype
style of faulting.

Z1.0
the depth to the shear-wave velocity horizon of 1.0(km/s).

regionflag
Regional attenuation flag, 0 = Global, 1 = China-Turkey, 2 = Italy-Japan.

basinflag
Basin Adjustments.

Value

A list will be return, including mag, Rjb, specT, period, lnY, sigma, iflag, Vs30, ftype, pga4nl, Z10, regionflag, basinflag, phi, tau.

Details

Boore, D. M., J. P. Stewart, E. Seyhan, and G. M. Atkinson (2014), NGA-west2 equations for predicting PGA, PGV, and 5 http://dx.doi.org/10.1193/070113EQS184M

Examples

BSSA14(6, 20, 0, 760, 0, 0.5, 0, 0)

#> $mag
#> [1] 6
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $Rjb
#> [1] 20
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $specT
#> [1] 0
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $period
#> [1] 0
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $lnY
#> [1] 4.578145
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $sigma
#> [1] 0.605122
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $iflag
#> [1] 0
#> 
#> $Vs30
#> [1] 760
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $ftype
#> [1] 0
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $pga4nl
#> [1] 0.09907733
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $Z1.0
#> [1] 0.5
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $regionflag
#> [1] 0
#> 
#> $basinflag
#> [1] 0
#> 
#> $phi
#> [1] 0.4950712
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $tau
#> [1] 0.3479613
#> attr(,"Csingle")
#> [1] TRUE
#> 

BSSA14(7, 20, 0, 760, 0, 0.5, 0, 0)

#> $mag
#> [1] 7
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $Rjb
#> [1] 20
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $specT
#> [1] 0
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $period
#> [1] 0
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $lnY
#> [1] 4.990961
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $sigma
#> [1] 0.605122
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $iflag
#> [1] 0
#> 
#> $Vs30
#> [1] 760
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $ftype
#> [1] 0
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $pga4nl
#> [1] 0.1497124
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $Z1.0
#> [1] 0.5
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $regionflag
#> [1] 0
#> 
#> $basinflag
#> [1] 0
#> 
#> $phi
#> [1] 0.4950712
#> attr(,"Csingle")
#> [1] TRUE
#> 
#> $tau
#> [1] 0.3479613
#> attr(,"Csingle")
#> [1] TRUE
#>